fixing some package order
