+++ app/src/main/res/layout/fragment_broadcaster_area.xml
@@ -5,7 +5,7 @@
     android:id="@+id/ConstraintLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Fragments.BroadcasterAreaFragment">
+    tools:context=".Fragments.PersonalArea.BroadcasterAreaFragment">
 
     <androidx.constraintlayout.widget.Guideline
         android:id="@+id/guideline2"
