+++ app/src/main/res/layout/fragment_broadcaster_profile.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Fragments.BroadcasterProfileFragment">
+    tools:context=".Fragments.PersonalArea.BroadcasterProfileFragment">
 
     <TextView
         android:id="@+id/textView3"
