+++ app/src/main/res/layout/fragment_broadcasting_contents.xml
@@ -5,7 +5,7 @@
     android:id="@+id/ConstraintLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Fragments.BroadcastingContentsFragment">
+    tools:context=".Fragments.PersonalArea.BroadcastingContentsFragment">
 
     <TextView
         android:id="@+id/textView5"
