+++ app/src/main/res/layout/fragment_personal_area.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Fragments.PersonalAreaFragment">
+    tools:context=".Fragments.PersonalArea.PersonalAreaFragment">
 
     <ImageView
         android:id="@+id/imageView"
