+++ app/src/main/res/layout/fragment_user_area.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Fragments.UserAreaFragment">
+    tools:context=".Fragments.PersonalArea.UserAreaFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:layout_width="0dp"
